+++ src/actions/clean.rs
@@ -4,8 +4,8 @@ use crate::{command, project::Project};
 
 /// Removes wasm folders, and runs `cargo clean`.
 pub fn clean_action(project: &Project) {
-    project.members.iter().for_each(|member| {
-        command::rm_dir(member.root.join("wasm"));
+    project.workspace_members.iter().for_each(|member_root| {
+        command::rm_dir(member_root.join("wasm"));
     });
 
     command::rm_dir(project.project_root().join("wasm"));